service_1/requirements.txt
flask
flask_sqlalchemy
flask_wtf
flask_testing
wtforms
pymysql
selenium
pytest
pytest-cov
requests
requests_mock